move packers to dedicated folder & add more XML comments to public nodes
